Extract league routes to own file
